+++ src/plone/restapi/serializer/dxcontent.py
@@ -11,6 +11,7 @@ from plone.restapi.batching import HypermediaBatch
 from plone.restapi.interfaces import IFieldSerializer
 from plone.restapi.interfaces import ISerializeToJson
 from plone.restapi.interfaces import ISerializeToJsonSummary
+from plone.restapi.search.handler import SearchHandler
 from plone.restapi.serializer.converters import json_compatible
 from plone.restapi.serializer.expansion import expandable_elements
 from plone.supermodel.utils import mergedTaggedValueDict
@@ -138,8 +139,23 @@ class SerializeFolderToJson(SerializeToJson):
         if batch.links:
             result['batching'] = batch.links
 
-        result['items'] = [
-            getMultiAdapter((brain, self.request), ISerializeToJsonSummary)()
-            for brain in batch
-        ]
+        if 'fullobjects' in self.request.form.keys():
+            handler = SearchHandler(self.context, self.request)
+            result['items'] = handler.search(
+                query={
+                    'fullobjects': 'true',
+                    'path': {
+                        'query': '/'.join(self.context.getPhysicalPath()),
+                        'depth': 1
+                    }
+                }).get('items')
+        else:
+            result['items'] = [
+                getMultiAdapter(
+                    (brain, self.request),
+                    ISerializeToJsonSummary
+                )()
+                for brain in batch
+            ]
+
         return result